requirements.txt
bottle
cheroot
RPi.GPIO
spidev
schedule
adafruit-blinka
adafruit-circuitpython-max31855
paho-mqtt
simple-pid
pyacaia
pygatt
pexpect
pillow
asyncio
bleak
asyncio-mqtt
